+++ RELEASE_NOTES.md
@@ -35,6 +35,12 @@ Logtalk compiler and runtime
 `synchronized` even when using backend Prolog compilers that don't support
 a compatible threads implementation.
 
+* CHANGED: An attempt to use a `::Closure` as a meta-argument in a call to a
+meta-predicate now throws a domain error. Previously, the corresponding `::/1`
+goal would only work for local meta-calls and would fail silently for calls in
+the context of the *sender*. The workaround is to call the `self/1` built-in
+method first and to pass to the meta-predicate a `Self::Closure` closure.
+
 * FIXED: Compilation of synchronized predicates when using backend Prolog
 compilers that don't support a compatible threads implementation.